Do not allow aria-owns to parent objects with no layout
If an object is unreachable by layout, it normally would not get
an AXObject created for it, because it cannot have parents reaching
to the root. However, aria-owns was breaking this logic, causing
situations where an AXObject could be created without a legal parent.

This CL adds a rule this disallows aria-owns from pointing to
elements that have no layout object. While it appears to be a bit
strong to disallow this, there were no clear use cases for doing it,
and the rule seems to provide a useful guide rail preventing
abnormal aria-owns corner cases from breaking Blink accessibility.

One consideration is elements that have no layout because they use
content-visibility:auto. However, these objects are recreated once
they gain layout, at which time the aria-owns relationship will be
valid.

Fixed: 1273502,1260750
Change-Id: I48718557a9ee1805a0fdd474e688f1dcd105fe0c