Add support for secondary views in WebXR
This change adds support for requesting secondary views in WebXR.
Secondary views can be requested by passing in 'secondary-views'
to requestSession. With this enabled, an arbitrary number of
secondary views can be exposed from the XR runtime on each frame.

This change also implements secondary views in the OpenXR backend,
supporting the first-person observer view configuration. Secondary views
in OpenXR are only exposed to WebXR only when they are active (such as
when the camera feed is being captured) and requested by WebXR.

Change-Id: Icb6a886e3f3b907f0c619bce5cf1f1481fb38a44